dev: move Term from cleaning to util
