+++ covsirphy/analysis/model_validator.py
@@ -4,7 +4,7 @@
 from itertools import chain
 import numpy as np
 import pandas as pd
-from covsirphy.cleaning.term import Term
+from covsirphy.util.term import Term
 from covsirphy.ode.mbase import ModelBase
 from covsirphy.cleaning.population import PopulationData
 from covsirphy.phase.phase_estimator import MPEstimator
